backend/requirements.txt
Django==4.1.5
psycopg2>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.2.0
requests==2.28.1
django-websocketclient==0.1.0
django-background-tasks-updated==1.2.7